requirements.txt
requests
schedule
python-telegram-bot==13.15
python-dotenv
numpy
